+++ src/javascript/binary/websocket_pages/resources/asset_index/asset_indexws.ui.js
@@ -1,5 +1,7 @@
 var showLoadingImage = require('../../../base/utility').showLoadingImage;
 var Table = require('../../../common_functions/attach_dom/table').Table;
+var AssetIndexData = require('./asset_indexws.data').AssetIndexData;
+var AssetIndex = require('../asset_indexws').AssetIndex;
 
 var AssetIndexUI = (function() {
     "use strict";
@@ -29,7 +31,10 @@ var AssetIndexUI = (function() {
         showLoadingImage($container);
 
         isFramed = (config && config.framed);
-        if (!assetIndex) AssetIndexData.sendRequest(!activeSymbols);
+        if (!assetIndex) {
+            initSocket();
+            AssetIndexData.sendRequest(!activeSymbols);
+        }
     };
 
     var populateTable = function() {
@@ -118,6 +123,28 @@ var AssetIndexUI = (function() {
         $submarketTable.find('thead').prepend($submarketHeader);
 
         return $submarketTable;
+    };
+
+    var initSocket = function() {
+        if (TradePage_Beta.is_trading_page()) return;
+        BinarySocket.init({
+            onmessage: function(msg) {
+                var response = JSON.parse(msg.data);
+                if (response) {
+                    responseHandler(response);
+                }
+            }
+        });
+    };
+
+    var responseHandler = function(response) {
+        var msg_type = response.msg_type;
+        if (msg_type === "asset_index") {
+            setAssetIndex(response);
+        }
+        else if (msg_type === "active_symbols") {
+            setActiveSymbols(response);
+        }
     };
 
     return {
