+++ src/javascript/binary/websocket_pages/trade/beta/analysis.js
@@ -1,3 +1,5 @@
+var AssetIndexUI = require('../../resources/asset_index/asset_indexws.ui').AssetIndexUI;
+
 /*
  * This file contains the code related to loading of trading page bottom analysis
  * content. It will contain jquery so as to compatible with old code and less rewrite
