+++ src/javascript/binary/websocket_pages/websocket_pjax.js
@@ -3,6 +3,7 @@ var Cashier = require('./cashier/cashier').Cashier;
 var ForwardWS = require('./cashier/deposit_withdraw_ws').ForwardWS;
 var PaymentAgentListWS = require('./cashier/payment_agent_listws').PaymentAgentListWS;
 var PaymentAgentWithdrawWS = require('./cashier/payment_agent_withdrawws').PaymentAgentWithdrawWS;
+var AssetIndexUI = require('./resources/asset_index/asset_indexws.ui').AssetIndexUI;
 
 pjax_config_page_require_auth("user/profit_table", function(){
     return {
